applied new core branch changes (not tested)
