+++ src/utils/testTextOutput.js
@@ -2,4 +2,8 @@
  * ? Purpose TBC
  * @param {Object} msg Discord message object
  */
-export default function TestTextOutput(msg) {}
+export default function TestTextOutput(msg) {
+	// let replyString = "";
+	// msg.reply(replyString);
+	// console.log(replyString);
+}
